realtime_pipeline/producer/requirements.txt
kafka-python
faker
